10-선형회귀_예시(2)를 통한 회귀분석 순서 및 프로젝트 가이드라인

+++ helper.py
@@ -284,7 +284,7 @@ def setCategory(df, fields=[], labelling=True):
                 for i, v in enumerate(cdf[field_name].dtypes.categories):
                     mydict[v] = i
                 
-                cdf[field_name] = cdf[field_name].map(mydict)
+                cdf[field_name] = cdf[field_name].map(mydict).astype(int)
 
     return cdf
